refactor: [move] application/service 패키지로 이동
